no stack (in env), mouse events, event capture, insert & hints modes
I was thinking that the stack wasn't useful so I was going to remove it entirely, but it turns out that it is useful, just not in the way I was originally thinking. I had originally thought that the mappings would reset the stack when they processed the events, but that seemed silly since they were state machines and didn't generally need the stack. After removing it I realized that to properly handle events that weren't mapped (as in a series of key strokes for example) I needed to capture them, sequester, and then replay if nothing mapped them. So I readded the stack as a variable in the processor itself. I may move it back out to the environment later if it turns out to be useful.

Added basic mouse events.

Events are generally captured now, except for the *down ones. There are a number of headaches with capturing and then trying to replay events. Not the least of which is that if I call preventDefault I can't undo it later. This still needs more work.

I've added rudimentary support for insert and hints modes. These also need more work...